TsWriter improvements - EPG time-zone/DST conversion/handling for ATSC, SCTE, MHW and OpenTV uses TOT/STT - MHW 1 general improvements; text and time decoding - MHW 2 support extension; episode/series details, channel categories, classifications - OpenTV and MHW EPG grabbers pass out text language - support more than one OpenTV channel category
...plus a few other minor refactorings and bug fixes

+++ DirectShowFilters/TsWriter/source/IGrabberEpgAtsc.h
@@ -45,7 +45,7 @@ DECLARE_INTERFACE_(IGrabberEpgAtsc, IGrabber)
                               unsigned short* sourceId,
                               unsigned short* eventId,
                               unsigned long long* startDateTime,
-                              unsigned short* duration,
+                              unsigned long* duration,
                               unsigned char* textCount,
                               unsigned long* audioLanguages,
                               unsigned char* audioLanguageCount,
